Split middleware from default library to enable front-end usage, require entrypoint /express.js to add middleware
