+++ lib/roles.js
@@ -1,8 +1,10 @@
 const _ = require('lodash');
 const jsonwebtoken = require('jsonwebtoken');
+
 const Role = require('./role');
 const Scope = require('./scope');
 const roleStringEquality = require('./role-string-equality');
+const jwtDecode = require('jwt-decode');
 
 class Roles {
 	constructor(...args) {
@@ -29,9 +31,13 @@ class Roles {
 	}
 
 	tryParseWebToken(token) {
-		const decoded = jsonwebtoken.decode(token);
-		if (!decoded) return false;
+		let decoded;
 
+		try {
+			decoded = jwtDecode(token);
+		} catch (e) {
+			return false;
+		}
 		if (decoded && decoded.roles && _.isArray(decoded.roles)) {
 			decoded.roles.forEach(element => this.parseElement(element));
 			return true;
